requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.7.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
celery==5.4.0
redis==5.2.1
httpx==0.28.1
python-json-logger==3.2.1
PyJWT==2.9.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
cryptography==44.0.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0
twilio==9.4.5
croniter==6.0.0
python-multipart==0.0.20
slowapi==0.1.9
